Remove location sharing item and CoreLocation
This dependency on CoreLocation adds the requirement to add permission description strings, which are required when submitting to Test Flight/App Store, to an app using `JSQMessagesViewController` even if it is not actually utilising `JSQLocationMediaItem`. Although this is not a general fix, it quick and dirty ad hoc fixup to avoid App Store rejections.

Also updating the project to Xcode 10.2 recommended settings.